NOVAS

Documentation for NOVAS.

NOVAS.cio_location

NOVAS.era

NOVAS.frame_tie

NOVAS.ira_equinox

NOVAS.nutation

NOVAS.precession

NOVAS.sidereal_time

NOVAS.tdb2tt

NOVAS.cio_location — Method

ciolocation(jdtdb)

Computes the location of the celestial intermediate origin (CIO) for a given Julian date, as a right ascension with respect to the true equinox of date.

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Accuracy=full: Either full or reduced accuracy

NOVAS.era — Method

era(jd_high)

This function returns the value of the earth rotation angle (theta in degrees) for a given UT1 Julian date.

Arguments

jd_high::Real: High-order part of UT1 Julian date

jd_low::Real: Low-order part of UT1 Julian date

NOVAS.frame_tie — Function

frame_tie(pos;direction)

Transforms a position vector from the dynamical reference system to ICRS, or vice versa.

Argumenets

pos::AbstractVector: Position vector, equitorial rectangular coordinates

direction::FrameTieDirection=dynamic2icrs

NOVAS.ira_equinox — Method

ira_equinox(jd_tdb)

Computes the intermediate right ascension of the equinox at the input Julian date using an analytical expression for the accumulated precession in right ascension. For the true equinox, the result is the equation of the origins.

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Accuracy=full: Either full or reduced accuracy

equinox::EquinoxType=mean_equinox: Either mean_equinox or true_equinox

NOVAS.nutation — Method

nutation(jd_tdb)

Nutates equitorial rectangular coordinates from mean equator and equinox of epoch to true equator or does the transformation in reverse given the flag direction.

Arguments

jd_tdb::Real: TDB Julian Date

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, referred to mean equator and equinox of epoch

Optional Arguments

accuracy::Accuracy=full: Either full or reduced accuracy

direction::NutationDirection=mean2true: Perform the transformation in either the mean2true or true2mean direction

NOVAS.precession — Method

precession(jd_tdb1,pos,jd_tdb2)

Precesses equatorial rectangular coordinates from one epoch to another. One of the two epochs must be J200.0.

Arguments

jd_tdb1::Real: TDB Julian Date of first epoch

jd_tdb2::Real: TDB Julian Date of second epoch

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, referred to mean dynamical equator and equinox of first epoch

NOVAS.sidereal_time — Method

sidereal_time(jd_ut1,0,delta_t)

Computes the Greenwich sidereal time, either mean or apparent, at Julian date jd_high + jd_low in hours

Arguments

jd_high::Real: high-order part of UT1 Julian date

jd_low::Real: low-order part of UT1 Julian date

delta_t::Real: Difference TT-UT1 at jd_high + jd_low in seconds

Optional arguments

gst_type::GSTType=mean_gst: Return results as mean (mean_gst) or apparent (apparent_gst) time

method::SiderealMethod=equinox: Computation method, CIO-based (CIO) or equinox-based (equinox)

accuracy::Accuracy=full: Either full or reduced accuracy

NOVAS.tdb2tt — Method

tdb2tt(tdb_jd)

Computes the Terrestrial Time (TT) or Terrestrial Dynamic Time (TDT) Julian date corresponding to a Barrycentric Dynamical Time (TDB) Julian Date)

Arguments

tdb_jd::Real: TDB Julian date