NOVAS

Documentation for NOVAS.

NOVAS.e_tilt

NOVAS.era

NOVAS.sidereal_time

NOVAS.tdb2tt

NOVAS.e_tilt — Method

e_tilt!(jd_tdb)

Computes quantities related to the orientation of the Earth's rotation axis at Julian date jd_tdb

Arguments

jd_tdb::Real: TDB Julian Date

accuracy=:full: Sets the accuracy level of :full or :reduced

Returns

(mobl,tobl,ee,dpsi,deps) where

mobl: Mean obliquity of the ecliptic in degrees

tobl: True obliquity of the ecliptic in degrees

ee: Equation of the equinoxes in seconds of time

dpsi: Nutation in longitude on arcseconds

deps: Nutation in obliquity in arcseconds

NOVAS.era — Method

era(jd_high)

This function returns the value of the earth rotation angle (theta in degrees) for a given UT1 Julian date.

Arguments

jd_high::Real: High-order part of UT1 Julian date

jd_low::Real: Low-order part of UT1 Julian date

NOVAS.sidereal_time — Method

sidereal_time(jd_ut1,0,delta_t)

Computes the Greenwich sidereal time, either mean or apparent, at Julian date jd_high + jd_low in hours

Arguments

jd_high::Real: high-order part of UT1 Julian date

jd_low::Real: low-order part of UT1 Julian date

delta_t::Real: Difference TT-UT1 at jd_high + jd_low in seconds

Optional arguments

gst_type: Return results as mean (:mean) or apparent (:apparent) time

method: Computation method, CIO-based (:CIO) or equinox-based (:equinox)

accuracy: Either :full or :reduced accuracy

NOVAS.tdb2tt — Method

tdb2tt(tdb_jd)

Computes the Terrestrial Time (TT) or Terrestrial Dynamic Time (TDT) Julian date corresponding to a Barrycentric Dynamical Time (TDB) Julian Date)

Arguments

tdb_jd::Real: TDB Julian date