API (Exported Functions)

NOVAS.OnSurface

NOVAS.cel_pole

NOVAS.cio_basis

NOVAS.cio_location

NOVAS.e_tilt

NOVAS.ee_ct

NOVAS.equ2hor

NOVAS.era

NOVAS.frame_tie

NOVAS.fund_args

NOVAS.iau2000a

NOVAS.ira_equinox

NOVAS.mean_obliq

NOVAS.norm_ang

NOVAS.nu2000k

NOVAS.nutation

NOVAS.nutation_angles

NOVAS.precession

NOVAS.refract

NOVAS.sidereal_time

NOVAS.spin

NOVAS.tdb2tt

NOVAS.ter2cel

NOVAS.wobble

NOVAS.cel_pole — Method

cel_pole(tjd)

Computes the celestial pole offsets for high-precision applications. Each set of offsets is a correction to the modeled position of the pole for a specific date, derived from observations and published by the IERS.

This function differs from the C version, where this returns the corrections instead of mutating global state.

Arguments

tjd::Real: TDB or TT Julian date for pole offsets

type::Symbol=:angular: Type of pole offset. :angular for corrections to angular coordinates of modeled pole referred to mean ecliptic of date, that is, delta-delta-psi and delta-delta-epsilon. :positional for corrections to components of modeled pole unit vector referred to GCRS axes, that is, dx and dy.

dople1: Value of celestial pole offset in first coordinate.

dpole2: Value of celestial pole offset is second coordinate.

NOVAS.cio_basis — Method

cio_basis(jd_tdb,ra_cio)

Computes the orthonomal basis vectors, with respect to the GCRS of the celestial intermediate system defined by the celestial intermediate pole and origin.

Arguments

jd_tdb::Real: TDB Julian Date

ra_cio::Real: Right ascension of the CIO at epoch (hours)

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

NOVAS.cio_location — Method

ciolocation(jdtdb)

Computes the location of the celestial intermediate origin (CIO) for a given Julian date, as a right ascension with respect to the true equinox of date.

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

NOVAS.e_tilt — Method

e_tilt(jd_tdb)

Computes quantities related to the orientation of the Earth's rotation axis at Julian date jd_tdb

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Symbol=:full: Sets the accuracy level of :full or :reduced

Returns

(mobl,tobl,ee,dpsi,deps) where

mobl: Mean obliquity of the ecliptic in degrees

tobl: True obliquity of the ecliptic in degrees

ee: Equation of the equinoxes in seconds of time

dpsi: Nutation in longitude on arcseconds

deps: Nutation in obliquity in arcseconds

NOVAS.ee_ct — Method

ee_ct(jd)

Computes the "complementary terms" of the equation of the equinoxes.

Arguments

jd_high::Real: High-order part of UT1 Julian date

jd_low::Real: Low-order part of UT1 Julian date

Optional Arguments

accuracy::Symbol=:full: Sets the accuracy level of :full or :reduced

NOVAS.equ2hor — Method

equ2hor(jd_ut1,delta_t,ra,dec)

This function transforms topocentric right ascension and declination to zenith distance and azimuth. It uses a method that properly accounts for polar motion, which is significant at the sub-arcsecond level. This function can also adjust coordinates for atmospheric refraction.

xp and yp are only needed if sub-arcsecond accuracy is needed.

Arguments

jd_ut1::Real: UT1 Julian Date

delta_t::Real: Difference TT-UT1 ast jd_ut1 in seconds

ra::Real: Topocentric right ascension in hours, referred to true equator

dec::Real: Topocentric declination in degrees, referred to true equatior

location::OnSurface: Structure containing observer's location details

Optional arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

xp::Real=0: x coordinate of the celestial intermediate pole in arcseconds

yp::Real=0: y coordinate of the celestial intermediate pole in arcseconds

ref_option::Symbol=:none: Whether to include refraction in the calculations.

Either :none, :standard to use standard atmospheric conditions, or :location to use the conditions included in location.

Returns

(zd,az,rar,decr) where

zd: Topocentric zenith distance in degrees

az: Topocentric azimuth (east of north) in degrees

rar: Topocentric RA of object of interest in hours

decr: Topocentric declination of object of interest in degrees

NOVAS.era — Function

era(jd_high)

This function returns the value of the earth rotation angle (theta in degrees) for a given UT1 Julian date.

Arguments

jd_high::Real: High-order part of UT1 Julian date

jd_low::Real=0.0: Low-order part of UT1 Julian date

NOVAS.frame_tie — Method

frame_tie(pos;direction)

Transforms a position vector from the dynamical reference system to ICRS, or vice versa.

Argumenets

pos::AbstractVector: Position vector, equitorial rectangular coordinates

direction::Symbol: Transformation direction, :dynamic2icrs or :icrs2dynamic

NOVAS.fund_args — Method

fund_args(t)

Computes the fundamental arguments (mean elements) of the Sun and Moon.

Arguments

-t::Real: TDB time in Julian centuries since J2000.0

Returns

[l,l′,F,D,Ω] where -l: Mean anomaly of the Moon -l′: Mean anomaly of the Sun -F: mean argument of the latitude of the Moon -D: Mean elongation of the Moon from the Sun -Ω Mean longitude of the Moon's ascending node

NOVAS.iau2000a — Function

iau2000a(jd)

Compute the forced nutation of the non-rigid earth based on the IAU 2000A nutation model.

Arguments

-jd_high::Real: High order part of the TT Julian date -jd_low::Real=0.0: Low order part of the TT Julian date

Return

(dpsi,deps) where -dpsi: Nutation (luni-solar + planetary) in longitude, in radians -deps: Nutation (luni-solar + planetary) in obliquity, in radians

NOVAS.ira_equinox — Method

ira_equinox(jd_tdb)

Computes the intermediate right ascension of the equinox at the input Julian date using an analytical expression for the accumulated precession in right ascension. For the true equinox, the result is the equation of the origins.

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

equinox::Symbol=:mean: Either :mean or :true

NOVAS.mean_obliq — Method

mean_obliq(jd_tdb)

Computes the mean obliquity of the ecliptic

NOVAS.norm_ang — Method

norm_ang(θ)

Normalizes an angle into range 0 <= θ <= 2π

NOVAS.nu2000k — Function

nu2000k(jd)

Compute the forced nutation of the non-rigid earth based on the NU2000K nutation model.

Arguments

-jd_high::Real: High order part of the TT Julian date -jd_low::Real=0.0: Low order part of the TT Julian date

Return

(dpsi,deps) where -dpsi: Nutation (luni-solar + planetary) in longitude, in radians -deps: Nutation (luni-solar + planetary) in obliquity, in radians

NOVAS.nutation — Method

nutation(jd_tdb,pos)

Nutates equitorial rectangular coordinates from mean equator and equinox of epoch to true equator or does the transformation in reverse given the flag direction.

Arguments

jd_tdb::Real: TDB Julian Date

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, referred to mean equator and equinox of epoch

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

direction::Symbol=:mean2true: Perform the transformation in either the :mean2true or :true2mean direction

NOVAS.nutation_angles — Method

nutation_angles(t)

Computes the values for nutation in longitude and nutation in obliquity for a given TDB julian date.

Aruments

t::Real: TDB time in Julian centuries since J2000.0

Optional Arguments

accuracy::Symbol=:full: Sets the accuracy level of :full or :reduced

Returns

(dpsi,deps) where

dpsi: Nutation in longitude in arcseconds

deps: Nutation in obliquity in arcseconds

NOVAS.precession — Method

precession(jd_tdb1,pos,jd_tdb2)

Precesses equatorial rectangular coordinates from one epoch to another. One of the two epochs must be J200.0.

Arguments

jd_tdb1::Real: TDB Julian Date of first epoch

jd_tdb2::Real: TDB Julian Date of second epoch

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, referred to mean dynamical equator and equinox of first epoch

NOVAS.refract — Method

refract(location,zd_obs)

Computes the atmospheric refraction in zenith distance. This is approximate for optical wavelengths.

Arguments

location::OnSurface: The observation location

zd_obs::Real: Observed zenith distance, in degrees

Optional Arguments

ref_option::Symbol=:standard: Either :standard to use standard atmospheric conditions, or :location to use the conditions included in location.

NOVAS.sidereal_time — Function

sidereal_time(jd_ut1,0,delta_t)

Computes the Greenwich sidereal time, either mean or apparent, at Julian date jd_high + jd_low in hours. See Chapter 5 in the NOVASC manual for information between equinox and CIO-based methods.

Arguments

jd_high::Real: high-order part of UT1 Julian date

jd_low::Real=0: low-order part of UT1 Julian date

delta_t::Real=0: Difference TT-UT1 at jd_high + jd_low in seconds

Optional arguments

gst_type::Symbol=:mean: Return results as mean (:mean) or apparent (:apparent) time

method::Symbol=:CIO: Computation method, :CIO-based or :equinox-based

accuracy::Symbol=:full: Either :full or :reduced accuracy

NOVAS.spin — Method

spin(angle,pos)

Transfroms a vector from one coordiate system to another with the same origin and axes rotated about the z-axis.

Arguments

angle::Real: Angle of rotation, positive counterclockwise when viewed from +z, in degrees

pos::AbstractVector: Position Vector

NOVAS.tdb2tt — Method

tdb2tt(tdb_jd)

Computes the Terrestrial Time (TT) or Terrestrial Dynamic Time (TDT) Julian date corresponding to a Barrycentric Dynamical Time (TDB) Julian Date)

Arguments

tdb_jd::Real: TDB Julian date

NOVAS.ter2cel — Method

ter2cel(jd_ut_high,jd_ut_low,delta_t,vec)

Rotates a vector from the terrestrial to the celestial system. Specifically, it transforms a vector in ITRS to GCRS by applying rotations for polar motion, Earth rotation, nutation, precession, and the dynamical-to-GCRS frame tie.

Arguments

jd_ut_high::Real: High order part of UT1 Julian date

jd_ut_low::Real: Low order part of UT1 Julian date

delta_t::Real: Value of TT-UT1 at the input date

vec::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, reference axes set by option

Optional Arguments

method::Symbol=:CIO: Computation method, :CIO-based or :equinox-based

accuracy::Symbol=:full: Either :full or :reduced accuracy

option::Symbol=:GRCS: Sets position vector to be referred to :GCRS axes or :equinox of date and equator

xp::Real=0: x coordinate of the celestial intermediate pole in arcseconds

yp::Real=0: y coordinate of the celestial intermediate pole in arcseconds

NOVAS.wobble — Method

wobble(tjd,xp,yp,pos)

Corrects a vector in the ITRS (rotating Earth-fixed system) for polar motion, and also corrects the longitude origin (by a tiny amount) to the Terrestrial Intermediate Origin (TIO). The ITRS vector is thereby transformed to the terrestrial intermediate system, based on the true (rotational) equator and TIO. Because the true equator is the plane orthogonal to the direction of the Celestial Intermediate Pole (CIP), the components of the output vector are referred to z and x axes toward the CIP and TIO, respectively.

Arguments

tjd::Real: TT or UT1 Julian Date

xp::Real: Conventionally-defined X coordinate of the CIP in arcseconds

yp::Real: Conventionally-defined Y coordinate of the CIP in arcseconds

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates

Optional Arguments

direction::Symbol=:itrs2terr: Either :itrs2terr or :terr2itrs

NOVAS.OnSurface — Type

OneSurface(lat,lon,h,temp,p)

Structure holding the observer's location on the surface of the Earth. Atmospheric parameters are optional and are only used by equ2hor in the calculation of refraction.

Arguments

latitude::Real: ITRS latitude in degrees; north positive

longitude::Real: ITRS longitude in degrees; east positive

height::Real: Height of observer (meters)

temperature::Real=0: Temperature (degrees Celsius)

pressure::Real=0: Atmospheric pressure (millibars)