API (Exported Functions)

NOVAS.cio_basis

NOVAS.cio_location

NOVAS.e_tilt

NOVAS.ee_ct

NOVAS.era

NOVAS.frame_tie

NOVAS.fund_args

NOVAS.iau2000a

NOVAS.ira_equinox

NOVAS.mean_obliq

NOVAS.nu2000k

NOVAS.nutation

NOVAS.nutation_angles

NOVAS.precession

NOVAS.sidereal_time

NOVAS.tdb2tt

NOVAS.cio_basis — Method

cio_basis(jd_tdb,ra_cio)

Computes the orthonomal basis vectors, with respect to the GCRS of the celestial intermediate system defined by the celestial intermediate pole and origin.

Arguments

jd_tdb::Real: TDB Julian Date

ra_cio::Real: Right ascension of the CIO at epoch (hours)

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

NOVAS.cio_location — Method

ciolocation(jdtdb)

Computes the location of the celestial intermediate origin (CIO) for a given Julian date, as a right ascension with respect to the true equinox of date.

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

NOVAS.e_tilt — Method

e_tilt(jd_tdb)

Computes quantities related to the orientation of the Earth's rotation axis at Julian date jd_tdb

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Symbol=:full: Sets the accuracy level of :full or :reduced

Returns

(mobl,tobl,ee,dpsi,deps) where

mobl: Mean obliquity of the ecliptic in degrees

tobl: True obliquity of the ecliptic in degrees

ee: Equation of the equinoxes in seconds of time

dpsi: Nutation in longitude on arcseconds

deps: Nutation in obliquity in arcseconds

NOVAS.ee_ct — Method

ee_ct(jd)

Computes the "complementary terms" of the equation of the equinoxes.

Arguments

jd_high::Real: High-order part of UT1 Julian date

jd_low::Real=0.0: Low-order part of UT1 Julian date

Optional Arguments

accuracy::Symbol=:full: Sets the accuracy level of :full or :reduced

NOVAS.era — Function

era(jd_high)

This function returns the value of the earth rotation angle (theta in degrees) for a given UT1 Julian date.

Arguments

jd_high::Real: High-order part of UT1 Julian date

jd_low::Real=0.0: Low-order part of UT1 Julian date

NOVAS.frame_tie — Method

frame_tie(pos;direction)

Transforms a position vector from the dynamical reference system to ICRS, or vice versa.

Argumenets

pos::AbstractVector: Position vector, equitorial rectangular coordinates

direction::Symbol: Transformation direction, :dynamic2icrs or :icrs2dynamic

NOVAS.fund_args — Method

fund_args(t)

Computes the fundamental arguments (mean elements) of the Sun and Moon.

Arguments

-t::Real: TDB time in Julian centuries since J2000.0

Returns

[l,l′,F,D,Ω] where -l: Mean anomaly of the Moon -l′: Mean anomaly of the Sun -F: mean argument of the latitude of the Moon -D: Mean elongation of the Moon from the Sun -Ω Mean longitude of the Moon's ascending node

NOVAS.iau2000a — Function

iau2000a(jd)

Compute the forced nutation of the non-rigid earth based on the IAU 2000A nutation model.

Arguments

-jd_high::Real: High order part of the TT Julian date -jd_low::Real=0.0: Low order part of the TT Julian date

Return

(dpsi,deps) where -dpsi: Nutation (luni-solar + planetary) in longitude, in radians -deps: Nutation (luni-solar + planetary) in obliquity, in radians

NOVAS.ira_equinox — Method

ira_equinox(jd_tdb)

Computes the intermediate right ascension of the equinox at the input Julian date using an analytical expression for the accumulated precession in right ascension. For the true equinox, the result is the equation of the origins.

Arguments

jd_tdb::Real: TDB Julian Date

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

equinox::Symbol=:mean: Either :mean or :true

NOVAS.mean_obliq — Method

mean_obliq(jd_tdb)

Computes the mean obliquity of the ecliptic

NOVAS.nu2000k — Function

nu2000k(jd)

Compute the forced nutation of the non-rigid earth based on the NU2000K nutation model.

Arguments

-jd_high::Real: High order part of the TT Julian date -jd_low::Real=0.0: Low order part of the TT Julian date

Return

(dpsi,deps) where -dpsi: Nutation (luni-solar + planetary) in longitude, in radians -deps: Nutation (luni-solar + planetary) in obliquity, in radians

NOVAS.nutation — Method

nutation(jd_tdb,pos)

Nutates equitorial rectangular coordinates from mean equator and equinox of epoch to true equator or does the transformation in reverse given the flag direction.

Arguments

jd_tdb::Real: TDB Julian Date

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, referred to mean equator and equinox of epoch

Optional Arguments

accuracy::Symbol=:full: Either :full or :reduced accuracy

direction::Symbol=:mean2true: Perform the transformation in either the :mean2true or :true2mean direction

NOVAS.nutation_angles — Method

nutation_angles(t)

Computes the values for nutation in longitude and nutation in obliquity for a given TDB julian date.

Aruments

t::Real: TDB time in Julian centuries since J2000.0

Optional Arguments

accuracy::Symbol=:full: Sets the accuracy level of :full or :reduced

Returns

(dpsi,deps) where

dpsi: Nutation in longitude in arcseconds

deps: Nutation in obliquity in arcseconds

NOVAS.precession — Method

precession(jd_tdb1,pos,jd_tdb2)

Precesses equatorial rectangular coordinates from one epoch to another. One of the two epochs must be J200.0.

Arguments

jd_tdb1::Real: TDB Julian Date of first epoch

jd_tdb2::Real: TDB Julian Date of second epoch

pos::AbstractVector: Position vector, geocentric equatorial rectangular coordinates, referred to mean dynamical equator and equinox of first epoch

NOVAS.sidereal_time — Function

sidereal_time(jd_ut1,0,delta_t)

Computes the Greenwich sidereal time, either mean or apparent, at Julian date jd_high + jd_low in hours

Arguments

jd_high::Real: high-order part of UT1 Julian date

jd_low::Real=0: low-order part of UT1 Julian date

delta_t::Real=0: Difference TT-UT1 at jd_high + jd_low in seconds

Optional arguments

gst_type::Symbol=:mean: Return results as mean (:mean) or apparent (:apparent) time

method::Symbol=:CIO: Computation method, CIO-based (:CIO) or equinox-based (:equinox)

accuracy::Symbol=:full: Either :full or :reduced accuracy

NOVAS.tdb2tt — Method

tdb2tt(tdb_jd)

Computes the Terrestrial Time (TT) or Terrestrial Dynamic Time (TDT) Julian date corresponding to a Barrycentric Dynamical Time (TDB) Julian Date)

Arguments

tdb_jd::Real: TDB Julian date